requirements.txt
absl-py==0.11.0
aiofiles==0.6.0
aiohttp==3.7.3
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
blis==0.4.1
bs4==0.0.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
cymem==2.0.4
Cython==0.29.14
DateTime==4.3
fastapi==0.63.0
Flask==1.1.2
Flask-Cors==3.0.10
fsspec==0.8.7
gast==0.3.3
gensim==3.8.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h11==0.12.0
h5py==2.10.0
HeapDict==1.0.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==0.17.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
loader==2017.9.11
locket==0.2.1
Logbook==1.5.3
logcc==2018.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
msgpack==1.0.2
multidict==5.0.2
murmurhash==1.0.4
nltk==3.5
npm==0.1.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
optional-django==0.1.0
packaging==20.4
pandas==1.1.4
partd==1.1.0
pendulum==2.1.2
Pillow==8.0.1
plac==1.1.3
pluggy==0.13.1
pony==0.7.13
preshed==3.0.4
prettytable==2.0.0
protobuf==3.14.0
psutil==5.8.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.1
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2020.4
pytzdata==2020.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
smart-open==3.0.0
sortedcontainers==2.3.0
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.4
starlette==0.13.6
tblib==1.7.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
termcc==2019.8.2
termcolor==1.1.0
terminaltables==3.1.0
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.52.0
typing-extensions==3.7.4.3
Unidecode==1.1.1
urllib3==1.26.2
uvicorn==0.13.4
wasabi==0.8.0
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
wordcloud==1.8.1
wrap==2018.3.1
wrapt==1.12.1
yarl==1.6.3
zict==2.0.0
zope.interface==5.2.0
